Rng validation works

+++ BeerApp.Dao/Models/Beer.cs
@@ -10,7 +10,7 @@ namespace BeerApp.Dao.Models
     public class Beer
     {
         [JsonProperty("id")]
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         [JsonProperty("name")]
         public string? Name { get; set; }
@@ -31,25 +31,25 @@ namespace BeerApp.Dao.Models
         public double Abv { get; set; }
 
         [JsonProperty("ibu")]
-        public long Ibu { get; set; }
+        public int Ibu { get; set; }
 
         [JsonProperty("target_fg")]
-        public long TargetFg { get; set; }
+        public int TargetFg { get; set; }
 
         [JsonProperty("target_og")]
-        public long TargetOg { get; set; }
+        public int TargetOg { get; set; }
 
         [JsonProperty("ebc")]
-        public long Ebc { get; set; }
+        public int Ebc { get; set; }
 
         [JsonProperty("srm")]
-        public long Srm { get; set; }
+        public int Srm { get; set; }
 
         [JsonProperty("ph")]
         public double Ph { get; set; }
 
         [JsonProperty("attenuation_level")]
-        public long AttenuationLevel { get; set; }
+        public int AttenuationLevel { get; set; }
     }
 
 }
